refactor: extract puzzle and viewport hooks
